style: ReviewView 전체 스크롤로 변경
