+++ POME/Global/Source/Base/BaseTableViewCell.swift
@@ -9,9 +9,10 @@ import UIKit
 
 class BaseTableViewCell: UITableViewCell {
     
-    let baseView = UIView().then{
-        $0.backgroundColor = .white
-    }
+    let baseView = UIView()
+//        .then{
+//        $0.backgroundColor = .white
+//    }
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
