+++ POME/Presentation/ViewControllers/Review/ReviewViewController.swift
@@ -22,16 +22,11 @@ class ReviewViewController: BaseTabViewController {
     
     var consumeList = [Reaction?](repeating: nil, count: 10){
         didSet{
-            mainView.consumeTableView.reloadData()
+            mainView.tableView.reloadData()
         }
     }
     
-    let mainView = ReviewView().then{
-        $0.firstEmotionFilter.filterButton.addTarget(self, action: #selector(filterButtonDidClicked), for: .touchUpInside)
-        $0.secondEmotionFilter.filterButton.addTarget(self, action: #selector(filterButtonDidClicked), for: .touchUpInside)
-        $0.reloadingButton.addTarget(self, action: #selector(reloadingButtonDidClicked), for: .touchUpInside)
-    }
-    
+    let mainView = ReviewView()
     var emptyView: ReviewEmptyView!
     
     //MARK: - Method
@@ -40,14 +35,16 @@ class ReviewViewController: BaseTabViewController {
         
         let sheet: EmotionFilterSheetViewController!
         
-        if(sender == mainView.firstEmotionFilter.filterButton){
+        guard let cell = mainView.tableView.cellForRow(at: [0,2]) as? ReviewFilterTableViewCell else { return }
+        //TODO: Tag 활용하는 방식으로 변경하기
+        if(sender == cell.firstEmotionFilter.filterButton){
             sheet = EmotionFilterSheetViewController.generateFirstEmotionFilterSheet()
         }else{
             sheet = EmotionFilterSheetViewController.generateSecondEmotionFilterSheet()
         }
         
         sheet.filterHandler = { emotion in
-            guard let filterView = sender.superview as? ReviewView.EmotionFilterView, let emotion = EmotionTag(rawValue: emotion) else { return }
+            guard let filterView = sender.superview as? ReviewFilterTableViewCell.EmotionFilterView, let emotion = EmotionTag(rawValue: emotion) else { return }
             filterView.setFilterSelectState(emotion: emotion)
         }
         
@@ -55,8 +52,11 @@ class ReviewViewController: BaseTabViewController {
     }
     
     @objc func reloadingButtonDidClicked(){
-        mainView.firstEmotionFilter.setFilterDefaultState()
-        mainView.secondEmotionFilter.setFilterDefaultState()
+        
+        guard let cell = mainView.tableView.cellForRow(at: [0,2]) as? ReviewFilterTableViewCell else { return }
+        
+        cell.firstEmotionFilter.setFilterDefaultState()
+        cell.secondEmotionFilter.setFilterDefaultState()
     }
     
     //MARK: - Override
@@ -73,12 +73,9 @@ class ReviewViewController: BaseTabViewController {
     }
     
     override func initialize(){
-        mainView.goalTagCollectionView.delegate = self
-        mainView.goalTagCollectionView.dataSource = self
-        
-        mainView.consumeTableView.separatorStyle = .none
-        mainView.consumeTableView.delegate = self
-        mainView.consumeTableView.dataSource = self
+        mainView.tableView.separatorStyle = .none
+        mainView.tableView.delegate = self
+        mainView.tableView.dataSource = self
     }
     
     override func topBtnDidClicked() {
@@ -95,7 +92,7 @@ extension ReviewViewController: UICollectionViewDelegate, UICollectionViewDataSo
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: GoalCategoryCollectionViewCell.cellIdentifier, for: indexPath) as? GoalCategoryCollectionViewCell else { return UICollectionViewCell() }
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: GoalTagCollectionViewCell.cellIdentifier, for: indexPath) as? GoalTagCollectionViewCell else { return UICollectionViewCell() }
 
         cell.goalCategoryLabel.text = goalCategoryList.isEmpty ? "···" : goalCategoryList[indexPath.row]
         
@@ -121,16 +118,16 @@ extension ReviewViewController: UICollectionViewDelegate, UICollectionViewDataSo
     }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        guard let cell = collectionView.cellForItem(at: indexPath) as? GoalCategoryCollectionViewCell else { return }
+        guard let cell = collectionView.cellForItem(at: indexPath) as? GoalTagCollectionViewCell else { return }
         if(selectedGoalCategory == 0 && indexPath.row != 0){
-            guard let cell = collectionView.cellForItem(at: [0,0]) as? GoalCategoryCollectionViewCell else { return }
+            guard let cell = collectionView.cellForItem(at: [0,0]) as? GoalTagCollectionViewCell else { return }
             cell.setUnselectState()
         }
         cell.setSelectState()
     }
 
     func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
-        guard let cell = collectionView.cellForItem(at: indexPath) as? GoalCategoryCollectionViewCell else { return }
+        guard let cell = collectionView.cellForItem(at: indexPath) as? GoalTagCollectionViewCell else { return }
         cell.setUnselectState()
     }
 }
@@ -142,7 +139,7 @@ extension ReviewViewController: UITableViewDelegate, UITableViewDataSource{
             emptyView = ReviewEmptyView()
             self.view.addSubview(emptyView)
             emptyView.snp.makeConstraints{
-                $0.top.leading.trailing.bottom.equalTo(mainView.consumeTableView)
+                $0.top.leading.trailing.bottom.equalTo(mainView.tableView)
             }
         }else if(emptyView != nil){
             emptyView.removeFromSuperview()
@@ -153,17 +150,35 @@ extension ReviewViewController: UITableViewDelegate, UITableViewDataSource{
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: ConsumeReviewTableViewCell.cellIdentifier, for: indexPath) as? ConsumeReviewTableViewCell else { return UITableViewCell() }
-        cell.mainView.firstEmotionTag.setTagInfo(when: .first, state: .happy)
-        cell.mainView.secondEmotionTag.setTagInfo(when: .second, state: .sad)
         
-        if let reaction = consumeList[indexPath.row] {
-            cell.mainView.myReactionBtn.setImage(reaction.defaultImage, for: .normal)
+        switch indexPath.row{
+        case 0:
+            guard let cell = tableView.dequeueReusableCell(withIdentifier: GoalTagsTableViewCell.cellIdentifier, for: indexPath) as? GoalTagsTableViewCell else { return UITableViewCell() }
+            cell.tagCollectionView.delegate = self
+            cell.tagCollectionView.dataSource = self
+            return cell
+        case 1:
+            guard let cell = tableView.dequeueReusableCell(withIdentifier: GoalDetailTableViewCell.cellIdentifier, for: indexPath) as? GoalDetailTableViewCell else { return UITableViewCell() }
+            return cell
+        case 2:
+            guard let cell = tableView.dequeueReusableCell(withIdentifier: ReviewFilterTableViewCell.cellIdentifier, for: indexPath) as? ReviewFilterTableViewCell else { return UITableViewCell() }
+            cell.firstEmotionFilter.filterButton.addTarget(self, action: #selector(filterButtonDidClicked), for: .touchUpInside)
+            cell.secondEmotionFilter.filterButton.addTarget(self, action: #selector(filterButtonDidClicked), for: .touchUpInside)
+            cell.reloadingButton.addTarget(self, action: #selector(reloadingButtonDidClicked), for: .touchUpInside)
+            
+            return cell
+        default:
+            guard let cell = tableView.dequeueReusableCell(withIdentifier: ConsumeReviewTableViewCell.cellIdentifier, for: indexPath) as? ConsumeReviewTableViewCell else { return UITableViewCell() }
+            cell.mainView.firstEmotionTag.setTagInfo(when: .first, state: .happy)
+            cell.mainView.secondEmotionTag.setTagInfo(when: .second, state: .sad)
+            
+            if let reaction = consumeList[indexPath.row] {
+                cell.mainView.myReactionBtn.setImage(reaction.defaultImage, for: .normal)
+            }
+            cell.mainView.setOthersReaction(count: indexPath.row)
+    //        cell.delegate = self
+            return cell
         }
-        cell.mainView.setOthersReaction(count: indexPath.row)
-//        cell.delegate = self
-        
-        return cell
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
